Ported from Mongodb to NeDB & updated webCache.

+++ server/src/webSockets/main.js
@@ -1,14 +1,11 @@
-const { ObjectId } = require('mongodb');
 const path = require('path');
 const { closeRooms } = require('../closeRooms/main.js');
-const io = require('../server/websocket.js');
-const { database } = require('../mongoServer/mongo.js');
+const io = require('../services/websocket.js');
 const { createWhatsappSession } = require('../createWhatsappSession/main.js');
 const fs = require('fs');
 
-// MongoDB Config
-const collectionConnectors = database.collection('connectors');
-const collectionSettings = database.collection('settings');
+//NeDB Imports
+const { dbSettings, dbConnectors } = require('../neDbServer/neDb.js')
 
 // Import global variables
 const activedSessions = require('../globalVariables/activedSessions.js');
@@ -20,30 +17,30 @@ let botId = '';
 // WebSocket connections
 io.on("connection", (socket) => {
 
-    // Inicial actived sessions
+    // Actived sessions
     socket.emit("active", {
         activedSessions
     });
 
-    // List connectos
+    // List connectors
     socket.on("listConnectors", () => {
         async function listConnectors() {
-            const dataConnectors = await collectionConnectors.find().toArray();
-            socket.emit("connectors", {
-                dataConnectors
+            dbConnectors.find({}, (err, dataConnectors) => {
+                if (err) {
+                    return;
+                } else {
+                    socket.emit("connectors", { dataConnectors });
+                    socket.emit("active", { activedSessions });
+                }
             });
-            socket.emit("active", {
-                activedSessions
-            });
-            return;
-        };
+        }
         listConnectors();
     });
 
     // Reload Whatsapp sessions
     socket.on("reloadSessions", (data) => {
         function loadExistingSessions() {
-            createWhatsappSession(data.sessionId, socket)
+            createWhatsappSession(data.sessionId, socket);
             closeRooms();
             setInterval(closeRooms, 60000);
             return;
@@ -62,16 +59,25 @@ io.on("connection", (socket) => {
             department: recData.data.department
         };
         async function addConnector() {
-            const existingConnector = await collectionConnectors.findOne({ number: recData.data.number });
-            if (existingConnector) {
-                socket.emit("error", { message: "thisRegistered" });
-                return;
-            } else {
-                await collectionConnectors.insertOne(dataConnector);
-                createWhatsappSession(dataConnector.number, socket);
-                return;
-            };
-        };
+            dbConnectors.findOne({ number: recData.data.number }, (err, existingConnector) => {
+                if (err) {
+                    socket.emit("error", { message: "internalError" });
+                    return;
+                }
+                if (existingConnector) {
+                    socket.emit("error", { message: "thisRegistered" });
+                    return;
+                } else {
+                    dbConnectors.insert(dataConnector, (err, newConnector) => {
+                        if (err) {
+                            socket.emit("error", { message: "internalError" });
+                            return;
+                        };
+                        createWhatsappSession(newConnector.number, socket);
+                    });
+                }
+            });
+        }
         addConnector();
     });
 
@@ -83,7 +89,9 @@ io.on("connection", (socket) => {
         socket.emit("active", {
             activedSessions
         });
-        await client.destroy();
+        if (client !== undefined || client !== null){
+            client.destroy();
+        }
         setTimeout(() => {
             activedSessions[data.number] = ''
             socket.emit("active", {
@@ -95,20 +103,25 @@ io.on("connection", (socket) => {
     // Delete connector data for database
     socket.on("deleteConnectors", (data) => {
         async function deleteConnectors() {
-            const deletedConnector = await collectionConnectors.deleteOne({ _id: new ObjectId(data.id) });
-            if (deletedConnector) {
-                const phoneNumber = data.number;
-                const sessionFolder = path.join(__dirname, ".wwebjs_auth", `session-${phoneNumber}`);
-                if (fs.existsSync(sessionFolder)) {
-                    fs.rmSync(sessionFolder, { recursive: true });
+            dbConnectors.remove({ _id: data.id }, {}, (err, numRemoved) => {
+                if (err) {
+                    console.error('Delete connector error:', err);
+                    return;
                 }
-            };
-            return;
-        };
+                if (numRemoved > 0) {
+                    const phoneNumber = data.number;
+                    const sessionFolder = path.join(__dirname, ".wwebjs_auth", `session-${phoneNumber}`);
+                    if (fs.existsSync(sessionFolder)) {
+                        fs.rmSync(sessionFolder, { recursive: true });
+                    }
+                }
+            });
+        }
         deleteConnectors();
     });
 
-    // Set Settings
+
+    // Set settings
     socket.on("saveSettings", async (recData) => {
         const dataSettings = {
             _id: recData.data._id,
@@ -117,94 +130,89 @@ io.on("connection", (socket) => {
             ip: recData.data.ip,
             minutes: recData.data.minutes
         };
-        const existingSetting = await collectionSettings.findOne({ _id: dataSettings._id });
-        if (existingSetting === null) {
-            await collectionSettings.insertOne(dataSettings);
+        dbSettings.update({ _id: dataSettings._id }, dataSettings, { upsert: true }, (err) => {
             return;
-        } else {
-            const filter = { _id: dataSettings._id };
-            const update = {
-                $set: {
-                    id: dataSettings.id,
-                    token: dataSettings.token,
-                    ip: dataSettings.ip,
-                    minutes: dataSettings.minutes
-                }
-            };
-            await collectionSettings.updateOne(filter, update);
-            return;
-        };
+        });
     });
 
-    // View Settings
+    // View settings
     socket.on("viewSetting", async () => {
-        const existingSetting = await collectionSettings.findOne({ _id: 1234567890 });
-        socket.emit("viewerSetting", (existingSetting));
+        dbSettings.findOne({ _id: 1234567890 }, (err, existingSetting) => {
+            if (err) {
+                return;
+            } else {
+                socket.emit("viewerSetting", existingSetting);
+            }
+        });
     });
 
     // Read and edit bot archives
     socket.on("botAndOptions", async (id) => {
-        botId = id
+        botId = id;
         if (botId === null) {
             return;
         }
-        const botPath = await collectionConnectors.findOne({ number: id });
-        if (botPath.botText !== null && botPath.botOptions !== null) {
-            socket.emit("textbot", (botPath.botText));
-            socket.emit("botoptions", (botPath.botOptions));
-        }
+        dbConnectors.findOne({ number: id }, (err, botPath) => {
+            if (err) {
+                return;
+            }
+            if (botPath && botPath.botText !== null && botPath.botOptions !== null) {
+                socket.emit("textbot", botPath.botText);
+                socket.emit("botoptions", botPath.botOptions);
+            }
+        });
     });
 
     // Save bot dialogs
     socket.once("insertText", async (value) => {
-        const success = await collectionConnectors.updateOne({ number: botId }, { $set: { botText: JSON.parse(value) } });
-        if (success) {
-
-            socket.emit("reload");
-            const client = sessions[botId];
-            activedSessions[botId] = 'loading';
-
-            socket.emit("active", {
-                activedSessions
-            });
-            
-            await client.destroy();
-
-            setTimeout(() => {
-                activedSessions[botId] = ''
+        dbConnectors.update({ number: botId }, { $set: { botText: JSON.parse(value) } }, {}, (err, numAffected) => {
+            if (err) {
+                return;
+            }
+            if (numAffected > 0) {
+                socket.emit("reload");
+                const client = sessions[botId];
+                activedSessions[botId] = 'loading';
                 socket.emit("active", {
                     activedSessions
                 });
-            }, 10000);
-
-            return;
-        };
+                if (client !== undefined || client !== null){
+                    client.destroy();
+                }
+                setTimeout(() => {
+                    activedSessions[botId] = '';
+                    socket.emit("active", {
+                        activedSessions
+                    });
+                }, 10000);
+            }
+        });
     });
 
     // Save bot options
     socket.once("insertOptions", async (value) => {
-        const success = await collectionConnectors.updateOne({ number: botId }, { $set: { botOptions: value } });
-        if (success) {
-
-            socket.emit("reload");
-            const client = sessions[botId];
-            activedSessions[botId] = 'loading';
-
-            socket.emit("active", {
-                activedSessions
-            });
-            
-            await client.destroy();
-
-            setTimeout(() => {
-                activedSessions[botId] = ''
+        dbConnectors.update({ number: botId }, { $set: { botOptions: value } }, {}, (err, numAffected) => {
+            if (err) {
+                return;
+            }
+            if (numAffected > 0) {
+                socket.emit("reload");
+                const client = sessions[botId];
+                activedSessions[botId] = 'loading';
                 socket.emit("active", {
                     activedSessions
                 });
-            }, 10000);
-
-            return;
-        };
+                if (client !== undefined || client !== null){
+                    client.destroy();
+                }
+                setTimeout(() => {
+                    activedSessions[botId] = '';
+                    socket.emit("active", {
+                        activedSessions
+                    });
+                }, 10000);
+            }
+        });
     });
 
 });
